ai_recon/requirements.txt
python-nmap==0.7.1
requests==2.31.0
rich==13.7.0
openai==1.3.7
python-dotenv==1.0.0
dnspython==2.4.2
